feat: calibration function
A calibration function gets a list of chessboard images and returns the camera matrix as well as the distortion coefficients. While calibrating, the images for which the chessboard corners are found are displayed with the corners drawn into them.

In the program flow the function gets called iff there are not already calibration parameters saved.